Split test/Test.hs (2294 lines) into thematic submodules
The monolithic test module was becoming unwieldy. Split into:

- Test.Helpers       (122 lines) — shared test helpers
- Test.CoreTests     (184 lines) — QuickCheck, unit, lifecycle, locale, i18n
- Test.WidgetTests   (576 lines) — UI, ScrollView, TextInput, Image, WebView, Styled, alignment, color
- Test.PlatformTests (967 lines) — Permission, SecureStorage, BLE, Dialog, AuthSession, Location, BottomSheet, Camera, HTTP
- Test.AppContextTests (288 lines) — Registration, AppContext, exception handler
- Test.ActionTests   (322 lines) — Action/OnChange, Widget Eq, incremental rendering

Test.hs is now 65 lines of wiring. All 190 tests pass unchanged.
