+++ test/Test/CoreTests.hs
@@ -1,0 +1,184 @@
+-- | Core framework tests: QuickCheck properties, unit tests, lifecycle,
+-- locale parsing, and i18n translation.
+module Test.CoreTests
+  ( qcProps
+  , unitTests
+  , lifecycleTests
+  , localeTests
+  , i18nTests
+  ) where
+
+import Test.Tasty
+import Test.Tasty.QuickCheck as QC
+import Test.Tasty.HUnit
+
+import Data.Either (isLeft)
+import Data.IORef (newIORef, readIORef, modifyIORef')
+import Data.List (sort)
+import Data.Map.Strict qualified as Map
+import Data.Text qualified as Text
+import Foreign.C.String (newCString, peekCString)
+import Foreign.Marshal.Alloc (free)
+import HaskellMobile (MobileApp(..), haskellGreet, haskellOnLifecycle)
+import HaskellMobile.Lifecycle
+  ( LifecycleEvent(..)
+  , loggingMobileContext
+  , lifecycleToInt
+  , lifecycleFromInt
+  , MobileContext(..)
+  )
+import HaskellMobile.Locale
+  ( Language(..)
+  , Locale(..)
+  , LocaleFailure(..)
+  , getSystemLocale
+  , parseLocale
+  , localeToText
+  )
+import HaskellMobile.I18n
+  ( Key(..)
+  , TranslateFailure(..)
+  , translate
+  )
+import Test.Helpers (oneTwoThree, testApp, withContext)
+
+qcProps :: TestTree
+qcProps = testGroup "(checked by QuickCheck)"
+  [ QC.testProperty "sort == sort . reverse" $
+      \list -> sort (list :: [Int]) == sort (reverse list)
+  , QC.testProperty "Fermat's little theorem" $
+      \x -> ((x :: Integer)^zeven  - x) `mod` zeven == 0
+  ]
+  where
+    zeven :: Integer
+    zeven = 7
+
+allEvents :: [LifecycleEvent]
+allEvents = [Create, Start, Resume, Pause, Stop, Destroy, LowMemory]
+
+unitTests :: TestTree
+unitTests = testGroup "Unit tests"
+  [ testCase "List comparison (different length)" $
+       oneTwoThree `compare` [1,2] @?= GT
+
+  -- the following test does not hold
+  , testCase "List comparison (same length)" $
+      oneTwoThree `compare` [1,2,3] @?= EQ
+  , testCase "haskellGreet returns correct greeting" $ do
+      cname <- newCString "World"
+      cresult <- haskellGreet cname
+      result <- peekCString cresult
+      free cresult
+      free cname
+      result @?= "Hello from Haskell, World!"
+  , testCase "haskellGreet with different input" $ do
+      cname <- newCString "Android"
+      cresult <- haskellGreet cname
+      result <- peekCString cresult
+      free cresult
+      free cname
+      result @?= "Hello from Haskell, Android!"
+  ]
+
+lifecycleTests :: TestTree
+lifecycleTests = testGroup "Lifecycle"
+  [ testCase "lifecycleToInt produces sequential codes 0-6" $
+      map lifecycleToInt allEvents @?= [0, 1, 2, 3, 4, 5, 6]
+  , testCase "lifecycleFromInt roundtrips for all events" $
+      mapM_ (\event ->
+        lifecycleFromInt (lifecycleToInt event) @?= Just event
+      ) allEvents
+  , testCase "lifecycleFromInt returns Nothing for unknown codes" $ do
+      lifecycleFromInt 7 @?= Nothing
+      lifecycleFromInt (-1) @?= Nothing
+      lifecycleFromInt 100 @?= Nothing
+  , testCase "callback receives dispatched event" $ do
+      ref <- newIORef ([] :: [LifecycleEvent])
+      withContext (\event -> modifyIORef' ref (++ [event])) $ \ctx ->
+        haskellOnLifecycle ctx 2
+      received <- readIORef ref
+      received @?= [Resume]
+  , testCase "unknown codes are silently ignored" $ do
+      ref <- newIORef (0 :: Int)
+      withContext (\_ -> modifyIORef' ref (+ 1)) $ \ctx -> do
+        haskellOnLifecycle ctx 99
+        haskellOnLifecycle ctx (-1)
+      count <- readIORef ref
+      count @?= 0
+  , testCase "all 7 event types received in order" $ do
+      ref <- newIORef ([] :: [LifecycleEvent])
+      withContext (\event -> modifyIORef' ref (++ [event])) $ \ctx ->
+        mapM_ (haskellOnLifecycle ctx . lifecycleToInt) allEvents
+      received <- readIORef ref
+      received @?= allEvents
+  , testCase "loggingMobileContext handles all events without throwing" $
+      mapM_ (onLifecycle loggingMobileContext) allEvents
+  , testCase "testApp context handles all events without throwing" $ do
+      app <- testApp
+      mapM_ (onLifecycle (maContext app)) allEvents
+  ]
+
+localeTests :: TestTree
+localeTests = testGroup "Locale"
+  [ testCase "parseLocale parses language-only tag" $
+      parseLocale "en" @?= Right (Locale En Nothing)
+  , testCase "parseLocale parses language-region tag" $
+      parseLocale "nl-NL" @?= Right (Locale Nl (Just "NL"))
+  , testCase "parseLocale handles underscore separator" $
+      parseLocale "en_US" @?= Right (Locale En (Just "US"))
+  , testCase "parseLocale normalises case" $
+      parseLocale "EN-us" @?= Right (Locale En (Just "US"))
+  , testCase "parseLocale accepts 3-digit region code" $
+      parseLocale "en-001" @?= Right (Locale En (Just "001"))
+  , testCase "parseLocale rejects empty tag" $
+      parseLocale "" @?= Left EmptyLocaleTag
+  , testCase "parseLocale rejects unknown language code" $
+      isLeft (parseLocale "xx") @?= True
+  , testCase "parseLocale rejects invalid language" $
+      isLeft (parseLocale "123") @?= True
+  , testCase "parseLocale rejects single-char language" $
+      isLeft (parseLocale "e") @?= True
+  , testCase "localeToText roundtrips" $ do
+      let locale = Locale Nl (Just "NL")
+      parseLocale (localeToText locale) @?= Right locale
+  , testCase "localeToText language-only" $
+      localeToText (Locale En Nothing) @?= "en"
+  , testCase "localeToText language-region" $
+      localeToText (Locale Nl (Just "NL")) @?= "nl-NL"
+  , testCase "getSystemLocale returns non-empty text" $ do
+      locale <- getSystemLocale
+      assertBool "locale should not be empty" (not (Text.null locale))
+  ]
+
+i18nTests :: TestTree
+i18nTests = testGroup "I18n"
+  [ testCase "translate finds exact locale match" $ do
+      let translations = Map.fromList
+            [ (Locale Nl (Just "NL"), Map.fromList [(Key "greeting", "Hallo")])
+            , (Locale En Nothing,     Map.fromList [(Key "greeting", "Hello")])
+            ]
+      translate translations (Locale Nl (Just "NL")) (Key "greeting") @?= Right "Hallo"
+  , testCase "translate falls back to language-only" $ do
+      let translations = Map.fromList
+            [ (Locale Nl Nothing, Map.fromList [(Key "greeting", "Hallo")])
+            ]
+      translate translations (Locale Nl (Just "BE")) (Key "greeting") @?= Right "Hallo"
+  , testCase "translate reports KeyNotFound for missing key" $ do
+      let translations = Map.fromList
+            [ (Locale En Nothing, Map.fromList [(Key "greeting", "Hello")])
+            ]
+      translate translations (Locale En Nothing) (Key "farewell")
+        @?= Left (KeyNotFound (Locale En Nothing) (Key "farewell"))
+  , testCase "translate reports LocaleNotFound for missing locale" $ do
+      let translations = Map.fromList
+            [ (Locale En Nothing, Map.fromList [(Key "greeting", "Hello")])
+            ]
+      translate translations (Locale Ja Nothing) (Key "greeting")
+        @?= Left (LocaleNotFound (Locale Ja Nothing))
+  , testCase "translate prefers exact match over fallback" $ do
+      let translations = Map.fromList
+            [ (Locale Nl Nothing,       Map.fromList [(Key "greeting", "Hallo")])
+            , (Locale Nl (Just "BE"),   Map.fromList [(Key "greeting", "Hoi")])
+            ]
+      translate translations (Locale Nl (Just "BE")) (Key "greeting") @?= Right "Hoi"
+  ]
